add image to README

+++ scripts/runner.py
@@ -22,41 +22,44 @@ def main(**args: Dict[str, Any]) -> None:
         os.mkdir(output_directory)
 
     for liver_name in liver_folders:
-        print("Processing " + liver_name)
+        print("\nProcessing " + liver_name)
         # Create liver output folder if it does not already exist
         liver_output_folder = os.path.join(output_directory, liver_name)
         if not os.path.exists(liver_output_folder):
             os.mkdir(liver_output_folder)
 
         liver_folder_path = os.path.join(images_directory, liver_name)
-        liver_files = os.listdir(liver_folder_path)
-        liver_images = []
-        for file in liver_files:
-            if file.endswith(".tiff") and magnification in file: # TODO: Use regex
-                liver_images.append(file)
+        if os.path.isdir(liver_folder_path) and not liver_folder_path.startswith('.'):
+            liver_files = os.listdir(liver_folder_path)
+            liver_images = []
+            for file in liver_files:
+                if file.endswith(".tiff") and magnification in file: # TODO: Use regex
+                    liver_images.append(file)
 
-        # Remove existing liver fat estimates
-        csv_file_path = os.path.join(liver_output_folder,
-                                     f'{liver_name}_fat_estimates.csv')
-        if os.path.exists(csv_file_path):
-            os.remove(csv_file_path)
+            # Remove existing liver fat estimates
+            csv_file_path = os.path.join(liver_output_folder,
+                                         f'{liver_name}_fat_estimates.csv')
+            if os.path.exists(csv_file_path):
+                os.remove(csv_file_path)
 
-        for image_name in liver_images:
-            print(image_name)
-            print("Beginning segmentation...")
-            segment_liver(images_directory, output_directory,
-                                          liver_name, image_name, is_frozen, ij)
-            print("Estimating steatosis...")
-            estimate_steatosis(images_directory, output_directory,
-                                          liver_name, image_name, is_frozen)
-            print(image_name + " complete!")
+            for image_name in liver_images:
+                print(image_name)
+                print("Beginning segmentation...")
+                segment_liver(images_directory, output_directory,
+                                              liver_name, image_name, is_frozen, ij)
+                print("Estimating steatosis...")
+                estimate_steatosis(images_directory, output_directory,
+                                              liver_name, image_name, is_frozen)
+                print(image_name + " complete!")
 
-        # Create slides for each liver
-        if pathologist_estimates:
-            powerpoint_save_path = os.path.join(liver_output_folder,
-                                                f'{liver_name}_slides.pptx')
-            helpers.make_powerpoint(images_directory, output_directory,
-                            pathologist_estimates, liver_name, powerpoint_save_path)
+            # Create slides for each liver
+            if pathologist_estimates:
+                powerpoint_save_path = os.path.join(liver_output_folder,
+                                                    f'{liver_name}_slides.pptx')
+                helpers.make_powerpoint(images_directory, output_directory,
+                                pathologist_estimates, liver_name, powerpoint_save_path)
+        else:
+            print(liver_folder_path + " is either hidden or not a directory. Skipping.")            
 
 
 if __name__ == "__main__":
